[ALPS07210288] cmdq: fixing ENUM compare with zero
Fixing coverity issue about ENUM compare with zero

MTK-Commit-Id: 641316b8bfd42cb845cbe9068f53c4995e6e6c15

CR-Id: ALPS07210288
Feature: cmdq_gce

+++ drivers/misc/mediatek/cmdq/v3/cmdq_mdp_common.c
@@ -1147,7 +1147,7 @@ static s32 cmdq_mdp_copy_cmd_to_task(struct cmdqRecStruct *handle,
 static void cmdq_mdp_store_debug(struct cmdqCommandStruct *desc,
 	struct cmdqRecStruct *handle)
 {
-	u32 len;
+	long len;
 
 	if (!desc->userDebugStr || !desc->userDebugStrLen)
 		return;
